AIResumeChecker/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
spacy==3.7.2
PyPDF2==3.0.1
python-docx==0.8.11
reportlab==4.0.4
scikit-learn==1.3.0
pandas==2.1.1
numpy==1.24.3
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7